Criação manual do grafo via CLI

+++ src/main/java/classes/Edge.java
@@ -3,9 +3,9 @@ package classes;
 public class Edge {
     Vertex origin;
     Vertex destiny;
-    float value;
+    int value;
 
-    Edge(Vertex origin, Vertex destiny, float value) {
+    Edge(Vertex origin, Vertex destiny, int value) {
         this.origin = origin;
         this.destiny = destiny;
         this.value = value;
